+++ src/main/java/com/tristankechlo/livingthings/config/entity/RaccoonConfig.java
@@ -44,7 +44,7 @@ public class RaccoonConfig {
 	@SuppressWarnings("unchecked")
 	public static List<SpawnData> getDefaultSpawns() {
 		List<SpawnData> spawns = new ArrayList<SpawnData>();
-		spawns.add(new SpawnData(20, 2, 6, Biomes.PLAINS, Biomes.FOREST, Biomes.BIRCH_FOREST,
+		spawns.add(new SpawnData(18, 2, 5, Biomes.PLAINS, Biomes.FOREST, Biomes.BIRCH_FOREST,
 				Biomes.OLD_GROWTH_BIRCH_FOREST, Biomes.SUNFLOWER_PLAINS, Biomes.DARK_FOREST, Biomes.FLOWER_FOREST));
 		return spawns;
 	}