+++ src/main/java/com/tristankechlo/livingthings/config/entity/SnailConfig.java
@@ -39,7 +39,7 @@ public class SnailConfig {
 		List<SpawnData> spawns = new ArrayList<SpawnData>();
 		spawns.add(new SpawnData(12, 3, 5, Biomes.PLAINS, Biomes.SUNFLOWER_PLAINS, Biomes.SWAMP, Biomes.FOREST,
 				Biomes.BIRCH_FOREST, Biomes.FLOWER_FOREST, Biomes.RIVER));
-		spawns.add(new SpawnData(15, 4, 6, Biomes.MUSHROOM_FIELDS));
+		spawns.add(new SpawnData(10, 4, 6, Biomes.MUSHROOM_FIELDS));
 		return spawns;
 	}