+++ src/main/java/com/tristankechlo/livingthings/entities/CrabEntity.java
@@ -12,6 +12,7 @@ import com.tristankechlo.livingthings.entities.misc.IMobVariants;
 import com.tristankechlo.livingthings.entities.misc.IScaleableMob;
 import com.tristankechlo.livingthings.init.ModEntityTypes;
 import com.tristankechlo.livingthings.misc.ILexiconEntry;
+import com.tristankechlo.livingthings.misc.LivingThingsTags;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
@@ -20,6 +21,8 @@ import net.minecraft.network.syncher.EntityDataSerializers;
 import net.minecraft.network.syncher.SynchedEntityData;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.tags.BlockTags;
+import net.minecraft.tags.Tag;
 import net.minecraft.util.TimeUtil;
 import net.minecraft.util.random.WeightedRandom;
 import net.minecraft.util.valueproviders.UniformInt;
@@ -48,8 +51,7 @@ import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.ServerLevelAccessor;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.pathfinder.BlockPathTypes;
 
 public class CrabEntity extends Animal implements IMobVariants, NeutralMob, IScaleableMob, ILexiconEntry {
@@ -60,6 +62,7 @@ public class CrabEntity extends Animal implements IMobVariants, NeutralMob, ISca
 			EntityDataSerializers.BYTE);
 	private static final ResourceLocation LEXICON_ENTRY = new ResourceLocation(LivingThings.MOD_ID,
 			"neutral_mobs/crab");
+	private static Tag<Block> spawnableOn = null;
 	private static final UniformInt rangedInteger = TimeUtil.rangeOfSeconds(20, 39);
 	private static final Ingredient BREEDING_ITEMS = Ingredient.of(Items.COD);
 	private int angerTime;
@@ -106,11 +109,13 @@ public class CrabEntity extends Animal implements IMobVariants, NeutralMob, ISca
 		this.targetSelector.addGoal(1, new ResetUniversalAngerTargetGoal<>(this, true));
 	}
 
-	public static boolean canCrabSpawn(EntityType<CrabEntity> animal, LevelAccessor world, MobSpawnType reason,
+	public static boolean checkCrabSpawnRules(EntityType<CrabEntity> animal, LevelAccessor world, MobSpawnType reason,
 			BlockPos pos, Random random) {
-		BlockState state = world.getBlockState(pos.below());
-		return (world.isWaterAt(pos))
-				|| (state.is(Blocks.GRASS_BLOCK) || state.is(Blocks.SAND) || state.is(Blocks.GRAVEL));
+		if (spawnableOn == null) {
+			spawnableOn = BlockTags.getAllTags().getTagOrEmpty(LivingThingsTags.CRAP_SPAWNABLE_ON);
+		}
+		return (world.isWaterAt(pos)) || spawnableOn.contains(world.getBlockState(pos.below()).getBlock())
+				&& isBrightEnoughToSpawn(world, pos);
 	}
 
 	@Override