+++ src/main/java/com/tristankechlo/livingthings/entities/GiraffeEntity.java
@@ -11,13 +11,17 @@ import com.tristankechlo.livingthings.entities.ai.BetterMeleeAttackGoal;
 import com.tristankechlo.livingthings.entities.misc.IMobVariants;
 import com.tristankechlo.livingthings.init.ModEntityTypes;
 import com.tristankechlo.livingthings.misc.ILexiconEntry;
+import com.tristankechlo.livingthings.misc.LivingThingsTags;
 
+import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.syncher.EntityDataAccessor;
 import net.minecraft.network.syncher.EntityDataSerializers;
 import net.minecraft.network.syncher.SynchedEntityData;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.tags.BlockTags;
+import net.minecraft.tags.Tag;
 import net.minecraft.util.TimeUtil;
 import net.minecraft.util.random.WeightedRandom;
 import net.minecraft.util.valueproviders.UniformInt;
@@ -46,7 +50,9 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.ServerLevelAccessor;
+import net.minecraft.world.level.block.Block;
 
 public class GiraffeEntity extends Animal implements NeutralMob, IMobVariants, ILexiconEntry {
 
@@ -56,11 +62,20 @@ public class GiraffeEntity extends Animal implements NeutralMob, IMobVariants, I
 			"neutral_mobs/giraffe");
 	private static final Ingredient BREEDING_ITEMS = Ingredient.of(Items.WHEAT);
 	private static final UniformInt rangedInteger = TimeUtil.rangeOfSeconds(20, 39);
+	private static Tag<Block> spawnableOn = null;
 	private int angerTime;
 	private UUID angerTarget;
 
 	public GiraffeEntity(EntityType<? extends GiraffeEntity> entityType, Level worldIn) {
 		super(entityType, worldIn);
+	}
+
+	public static boolean checkGiraffeSpawnRules(EntityType<GiraffeEntity> animal, LevelAccessor world,
+			MobSpawnType reason, BlockPos pos, Random random) {
+		if (spawnableOn == null) {
+			spawnableOn = BlockTags.getAllTags().getTagOrEmpty(LivingThingsTags.GIRAFFE_SPAWNABLE_ON);
+		}
+		return spawnableOn.contains(world.getBlockState(pos.below()).getBlock()) && isBrightEnoughToSpawn(world, pos);
 	}
 
 	@Override