+++ src/main/java/com/tristankechlo/livingthings/entities/SharkEntity.java
@@ -7,11 +7,14 @@ import com.tristankechlo.livingthings.LivingThings;
 import com.tristankechlo.livingthings.config.LivingThingsConfig;
 import com.tristankechlo.livingthings.entities.ai.BetterMeleeAttackGoal;
 import com.tristankechlo.livingthings.misc.ILexiconEntry;
+import com.tristankechlo.livingthings.misc.LivingThingsTags;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.tags.FluidTags;
+import net.minecraft.tags.Tag;
 import net.minecraft.util.TimeUtil;
 import net.minecraft.util.valueproviders.UniformInt;
 import net.minecraft.world.entity.EntityType;
@@ -37,7 +40,7 @@ import net.minecraft.world.entity.animal.WaterAnimal;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
-import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.material.Fluid;
 import net.minecraft.world.level.pathfinder.BlockPathTypes;
 import net.minecraft.world.phys.Vec3;
 
@@ -46,6 +49,7 @@ public class SharkEntity extends WaterAnimal implements NeutralMob, ILexiconEntr
 	private static final ResourceLocation LEXICON_ENTRY = new ResourceLocation(LivingThings.MOD_ID,
 			"hostile_mobs/shark");
 	private static final UniformInt rangedInteger = TimeUtil.rangeOfSeconds(20, 39);
+	private static Tag<Fluid> spawnableOn = null;
 	private int angerTime;
 	private UUID angerTarget;
 
@@ -146,9 +150,13 @@ public class SharkEntity extends WaterAnimal implements NeutralMob, ILexiconEntr
 		}
 	}
 
-	public static boolean canSharkSpawn(EntityType<SharkEntity> entity, LevelAccessor world, MobSpawnType reason,
+	public static boolean checkSharkSpawnRules(EntityType<SharkEntity> entity, LevelAccessor world, MobSpawnType reason,
 			BlockPos pos, Random random) {
-		return world.getBlockState(pos).is(Blocks.WATER) && world.getBlockState(pos.above()).is(Blocks.WATER);
+		if (spawnableOn == null) {
+			spawnableOn = FluidTags.getAllTags().getTagOrEmpty(LivingThingsTags.SHARK_SPAWNABLE_ON);
+		}
+		return spawnableOn.contains(world.getFluidState(pos).getType())
+				&& spawnableOn.contains(world.getFluidState(pos.above()).getType());
 	}
 
 	@Override