+++ src/main/java/com/tristankechlo/livingthings/entities/ShroomieEntity.java
@@ -8,6 +8,7 @@ import com.tristankechlo.livingthings.entities.ai.ShroomiePlantMushroomGoal;
 import com.tristankechlo.livingthings.entities.misc.IMobVariants;
 import com.tristankechlo.livingthings.init.ModEntityTypes;
 import com.tristankechlo.livingthings.misc.ILexiconEntry;
+import com.tristankechlo.livingthings.misc.LivingThingsTags;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
@@ -16,6 +17,8 @@ import net.minecraft.network.syncher.EntityDataSerializers;
 import net.minecraft.network.syncher.SynchedEntityData;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.tags.BlockTags;
+import net.minecraft.tags.Tag;
 import net.minecraft.util.TimeUtil;
 import net.minecraft.util.valueproviders.UniformInt;
 import net.minecraft.world.DifficultyInstance;
@@ -46,8 +49,7 @@ import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.ServerLevelAccessor;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.Block;
 
 public class ShroomieEntity extends Animal implements ILexiconEntry, IMobVariants {
 
@@ -57,12 +59,21 @@ public class ShroomieEntity extends Animal implements ILexiconEntry, IMobVariant
 			EntityDataSerializers.BYTE);
 	private static final Ingredient BREEDING_ITEMS = Ingredient.of(Items.WHEAT);
 	private static final UniformInt RANGED_INTEGER = TimeUtil.rangeOfSeconds(30, 60);
+	private static Tag<Block> spawnableOn = null;
 	private boolean canPlantMushroom;
 	private int mushroomCooldown;
 
 	public ShroomieEntity(EntityType<? extends ShroomieEntity> entityType, Level world) {
 		super(entityType, world);
 		canPlantMushroom = false;
+	}
+
+	public static boolean checkShroomieSpawnRules(EntityType<ShroomieEntity> animal, LevelAccessor world,
+			MobSpawnType reason, BlockPos pos, Random random) {
+		if (spawnableOn == null) {
+			spawnableOn = BlockTags.getAllTags().getTagOrEmpty(LivingThingsTags.SHROOMIE_SPAWNABLE_ON);
+		}
+		return spawnableOn.contains(world.getBlockState(pos.below()).getBlock()) && isBrightEnoughToSpawn(world, pos);
 	}
 
 	@Override
@@ -158,12 +169,6 @@ public class ShroomieEntity extends Animal implements ILexiconEntry, IMobVariant
 	public static AttributeSupplier.Builder createAttributes() {
 		return Mob.createMobAttributes().add(Attributes.MAX_HEALTH, LivingThingsConfig.SHROOMIE.health.get())
 				.add(Attributes.MOVEMENT_SPEED, LivingThingsConfig.SHROOMIE.speed.get());
-	}
-
-	public static boolean canShroomieSpawn(EntityType<? extends Animal> type, LevelAccessor world, MobSpawnType reason,
-			BlockPos pos, Random random) {
-		BlockState state = world.getBlockState(pos.below());
-		return (state.is(Blocks.GRASS_BLOCK) || state.is(Blocks.MYCELIUM)) && world.getRawBrightness(pos, 0) > 7;
 	}
 
 	@Override